Refactor message storage
