+++ src/actions.js
@@ -2,5 +2,5 @@ export * from './app/appActions';
 export * from './account/accountActions';
 export * from './events/eventActions';
 export * from './nav/navActions';
-export * from './stream/streamActions';
+export * from './message-list/messagesActions';
 export * from './userlist/userListActions';
