+++ src/constants.js
@@ -29,7 +29,6 @@ export const EVENT_UPDATE_MESSAGE_FLAGS = 'EVENT_UPDATE_MESSAGE_FLAGS';
 
 export const STREAM_FETCHING_MESSAGES = 'STREAM_FETCHING_MESSAGES';
 export const STREAM_FETCHED_MESSAGES = 'STREAM_FETCHED_MESSAGES';
-export const STREAM_FETCHING_FAILED = 'STREAM_FETCHING_FAILED';
 export const STREAM_SET_MESSAGES = 'STREAM_SET_MESSAGES';
 
 export const PRESENCE_RESPONSE = 'PRESENCE_RESPONSE';
